Fix to an undefined symbol issue for the Teuchos_BLAS class on windows
Since the BLAS class has specializations, which must be dllexported when
building shared on windows, a work around is needed so that the symbols will be
defined for the windows shared build. Explicitly instantiating the needed BLAS
instantiations is required to provide the symbols in the teuchos dll and fix the
issue.

On windows when a class is dllexported all of its public base classes are also
implicitly dllexported as well. This is a good thing generally, but in the case
of templated base classes has an interesting side effect. The base class
necessarily has to have all of its public symbols exported and there by defined
in the dll. This is a problem when the expected behavior is to have the compiler
generate the code.

Normally templates don't need the dllexport because they are generated in the
translation unit they are needed in and not put in a library. However, in this
case the fact that the specializations had to be dllexported means that all
instances of the involved templates have to be fully instantiated and exported
as well.