Reverting to using old scaling method.
I can't explain it, but using the Tpetra leftscale gives wrong eigenvalues.
Writing the matrix to disk reveals nothing -- it looks fine.
It's as if somehow the apply hasn't gotten the message that the matrix is scaled.

Convergence is back to its normal pleasing self.

Also changed "->" to "." in Utils::Write.