Enabling scaling and tranpose when matrix is really an Epetra_CrsMatrix.
Scaling works fine now for Epetra and Tpetra.
Transpose is a work in progress.  The matrix matrix multiply later segfaults when I do an explicit tranpose of what's really
an Epetra_CrsMatrix.  Of all the places it might happen, it's in the call to isFillComplete()!

Right now, Quaxel should just use the implicit transpose. (See ScalingTest for how to do this.)