Putting in a transpose for Epetra that works.
Put in a simple tranpose like that for Tpetra.  Downside is that it does insertion one matrix element at a time.

Why am I doing this?
Something is inconsistent between the transposed matrix and the matrix matrix
multiply.  In the case of Tpetra, we thought this was because the transposer produced a matrix whos row map wasn't 1-1.
However, in Epetra::RowMatrixTransposer, there are comments that lead me to believe the rowmap is in fact 1-1.

Also changed vector update in Hierarchy::Iterate, as Cthulhu doesn't allow this for Epetra objects, apparently because
Epetra doesn't.  You'll get an error like
"Throw test that evaluated to true: (alpha != 1) || (beta != 0)"
in Cthulhu_EpetraCrsMatrix.hpp.