Fixed PyTrilinos bugs related to storing objects as RCPs
I upgraded to SWIG version 2.0.4 and this introduced several warnings
about derived classes that were stored as smart pointers, but whose
base classes were not.  Subsequent testing revealed 25 segmentation
faults from the 71 PyTrilinos tests.  This commit fixes the base
classes to be handled with smart pointers (RCPs) as well, and it fixed
the segmentation faults.