Fixing Intel 12.0.4 SERIAL_DEBUG build of Trilinos packages (ticket:2241)
Here is what I did:

(*) Disabling Sundance and Mesquite since they are failing and not needed for CASL.

(*) Going from full debug build to optimized compiler options and Trilinos_ENABLE_DEBUG=TRUE.  The boating of the object code with template names must have been killing this Intel compiler.  We don't care to test if the compiler can handle a large amount of template code in a debug build.  More could be say but I will not here.  What we have is better than nothing.  We are getting around a deficent compiler and it does the testing we need to do.

This should fix the Thyra build errors and it removed Sundance and Mesquite so they will not show up.  After this very few failings tests should be left (e.g. Rythmos, Stokhos, etc.).  I will deal with those next.