Disable netcdf on windows builds
disabling netcdf since it is needed by seacas which is now being enabled by
default on windows. I don't know if we will eventually add netcdf to windows,
but I went with disabling the tpl instead of the SEACASExodus package since it
is conceivable that we will add netcdf at some point and that is the only known
issue preventing exodus from being enabled currently. It is likely that if we
later add netcdf that we would forget to re-enable exodus so this is slightly
safe down the road.

I also noticed that the configure option saying where to find boost's include
dirs was misspelled. I fixed that as well since we do currently have an
installation of boost on kalli.