More changes to support ParMETIS 4_0.
George changed all integers in interface to idx_t (which is defined to be
int32_t or int64_t by the user), even integers that clearly will always be
integers (e.g., number of constraints, number of geometric dimensions).
I had to make changes to the arguments of ParMETIS calls to pass the correct
types to the functions.

Code now runs with 64-bit ParMETIS 4_0.  Have to check the answers.