Added parameter list option for passing in the output stream.
Now you can pass the stratimikos output stream in through the "Output
Stream" parameter of the "NOX Stratimikos Options" sublist.  This is
convenient for redirecting all of the stratimikos (i.e., you want it,
but maybe you don't want it printed to the screen, like in a test).  To
get this to work, I had to interchange the order of parsing the sublist
and setting up stratimikos.