Created "unit" tests for SG analysis, coupled
I moved the SG analysis and coupled problems into the unit tests and
tested their results.  I also added some infrastructure to eliminate
almost all of the screen output unless it is requested (either on the
command line or if a test fails).  I say almost because it doesn't
appear possible to capture all of the Dakota output.