Fix bug in Epetra CrsMatrix
In bug 5370 Martin Kronbichler reported an issue with the Epetra CrsMatrix
having a memory error due to trying to delete memory that wasn't allocated in
the way expected in all cases. He provided a fix which was to change where the
check for a static profile was done such that it separate from the allocation
for each row.

Build/Test Cases Summary
Enabled Packages: Epetra
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=778,notpassed=0 (490.79 min)
1) SERIAL_RELEASE => passed: passed=751,notpassed=0 (230.25 min)