Remove Epetra_MpiSmpComm from hdf5 parrallel calls
There is an odd issue with allowing the Epetra_MpiSmpComm to be used as a
potential communicator for parallel hdf5. An odd error regarding an undefined
symbol to the typeinfo of Epetra_MpiSmpComm comes up when building in shared
only. This class is deprecated so I am just removing if from this code which
does remove the error from the build, but doesn't really fix it.