Update installation testing script for changes to exclude system
Updated the installation test scripts to now explicitly enable all of the
packages that we want to perform installation testing. I started with the list
of all released packages then explicitly disabled the packages that we are not
currently testing in installation testing. This is to make it very obvious what
the differences are between the release tarball and the installation testing.
Some of the packages that are disabled I don't know what the reasoning was,
however, I wanted to make sure that we get the exact same set of packages for
installation tomorrow as we did today. Each disabled package can then be
revisited later to determine if it can be enabled for installation testing or
not.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Packages
0) MPI_DEBUG => passed: passed=1299,notpassed=0 (102.34 min)
1) SERIAL_RELEASE => passed: passed=1313,notpassed=0 (45.91 min)
Other local commits for this build/test group: 410f91e, 973ae9a