MueLu: add a demo amalgamation routine to CoalesceDropFactory
not used yet. just for demonstration. only for constant block size, however
all things that have to be changed for variable block size are marked within
the code. The changes for variable block size are minimal.
TODO: The member variables for block size within CoalesceDropFactory are NOT used.
Input: unamalgamated Operator
Output: graph of amalgamated matrix