Disable BinUtils for Zoltan only tests
The zoltan only test builds on godel are explicitly C only builds and this is
causing an issue with the BinUtils tpl. The BinUtils tpl has some configure time
checks to make sure needed features are available. Those tests are apparently
written in c++. The recent update to our update script for getting a cmake
version for nightly testing is now pulling in a version of cmake that considers
trying to perform a configure time test that is written in a language that is
not enabled as an error. It looks like this was silently failing before. which
would disable the BinUtils tpl. For now I have just diabled the BinUtils tpl for
these builds, however, this is a temporary solution. We need to do something
reasonable in these cases where configure time tests request a language that is
not enabled by default.