Putting quotes around argument ${Kokkos_ENABLE_HWLOC} which can be empty.
CMake news flash: If you call a CMake macro or function that takes four
arguments such as:

TRIBITS_ADD_OPTION_AND_DEFINE(
  Kokkos_ENABLE_DeviceNUMA
  HAVE_KOKKOS_DEVICENUMA
  "Enable arrays using NUMA Device."
  ${Kokkos_ENABLE_HWLOC}
  )

and the varible Kokkos_ENABLE_HWLOC is undefined or is defined and is empty,
then CMake thinks you are only passing in three arguments.  This gives a CMake
error.

This defect (in Kokkos, not in the Tribits system) was exposed by configuring
with -DTrilinos_ENABLE_Kokkos=ON -DKokkos_ENABLE_Array=ON.

The fact that Kokkos_ENABLE_HWLOC is empty may or may not be a defect in
Kokkos but in general, a valid boolean mode we support in the Tribits system
is empty "" and CMake code needs to use quotes to account for that.

Build/Test Cases Summary
Enabled Packages: Kokkos
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
0) MPI_DEBUG => passed: passed=10,notpassed=0 (0.59 min)
1) SERIAL_RELEASE => passed: passed=10,notpassed=0 (0.57 min)
2) MPI_DEBUG_SS => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
3) SERIAL_RELEASE_SS => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)