Fix continuous integration build ctest driver error.
CMake 2.8.0 (which trilinos-test2 currently runs) does not have
the CMAKE_CURRENT_LIST_DIR dynamically scoped variable. This was
being checked for and set, but since it was placed in an if block
that checked if the variable existed, a parent scope that set
that variable would place it in static scope, so it would
not dynamically change within an include (which is the 2.8.4
behavior).

This commit removes the if block to make sure it is unconditionally
set.