Change Claps to experimental code
Claps had a new dependency on METIS added. We do not have METIS on most of our
build machines and zoltan also can use METIS, and potentially a different
version than what claps needs. I am setting claps to experimental with the
blessing of Clark and the Sierra/SD team. Claps is not a released package and
the primary users of it do not use our build system. To prevent confusion if
someone wants to enable claps I did add METIS as a dependency of claps.