Adding verboseUnitTests() function and test
I added the function Teuchos::UnitTestRepository::verboseUnitTest() so that
unit tests can determine if they are in verbose mode or not.  This is not
needed or desired for C++ code that uses std::ostream but if the unit tests
drive other lanuages like C or Fortran, then it is nice to be able to allow
the unit test to know if it is verbose or not so it can know to print output
or not.  This will not make for verbose output in the case of a failure but at
least it is better than always printing when you don't want to.

I added this for driving Fortran unit tests from C++ (for ForTrilinos work).

Build/Test Cases Summary
Enabled Packages: Teuchos
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
0) MPI_DEBUG => passed: passed=81,notpassed=0 (3.54 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
Other local commits for this build/test group: 3ffe8fd, 6ffefdb