stk_mesh: address a performance hot-spot that was revealed by an adagio use-case
When a problem has many mesh-parts, and when a field is registered on many parts,
an n-squared operation was being performed. (n is num-parts)
This checkin removes the n-squared operation for field-registration. Depending on
the order in which part-subsetting occurs (before or after field-registration)
there could still be another hot-spot. This scenario will be checked and corrected next.

(From Sierra)