Fix issue with tarball creation.
Will Dicharry noticed that make package_source wasn't working on the development
branch. It turns out that the call back for creating a package wasn't being
called because the wrong repo variable was being sent to the call back. The repo
directory was being sent, which evaluated to "." for Trilinos. The repo name is
what needed to be sent instead.

Interestingly this was not caught by the nightly installation testing we run
which does create a tarball. Apparently cmake keeps around files that describe
how to create a tarball which were not being deleted before each run. These old
files were still being used to create the tarball for the installation testing
which allowed it to continue to make a working tarball each night. This will be
addressed in a later commit.