MueLu: updated SetFactoryManager
I changed the interface of SetFactoryManager to use an RCP on the Level because
it's important that the level remains valid until the destruction of the SetFactoryManager object (otherwise -> segfault).
I saw this because in Hierarchy::Setup() (after d22ca73f8d9), SetFactoryManager is used on nextLevel and nextLevel might be deleted.

I also limited the scope of the current hack (9923b2b5c) to allow Tobias' use-cases to work again.