Fix issues with nightly installation testing
The script that creates the tarball for installation testing was using the wrong
version of the tarball for installation testing. There was still a 10.9 tarball
lying around and it was choosing that over the correct 10.11 tarball. To fix
this the script now removes the configure directory before attempting to
configure every night. This should also eliminate the possibility of the other
issue we had recently where the make package_source target wasn't working, but
the installation testing still worked since cmake kept around enough information
to continue to build an old version of the tarball. Starting with a completely
clean configure is just safer too.