Adding transfer factory for multivectors.
Most common use case is for transferring vectors from fine to coarse level via RAPFactory.
You can specify the name of the MultiVector, as well as the factory that should generate the transfer op.
Has a unit test, but not enabled in cmake for the moment.
Should probably be moved to another directory eventually, probably MueCentral.