Framework/STK relation integration fix
A corner case was exposed by a fuego test that caused relation
ordering to become invalid. As a temporary fix, we've changed the
find_relation method to use a relation comparator that does not take
entity id into account. Thw downside is that this leaves auxiliary
relations in a potentially unsorted state which has unknown
consequences.

(From Sierra)