make the end-of-run gathering of timer data more scalable.
a thermal-fluids run on 65000 processors failed at the end of the run when
trying to gather timing data from all processors.
This checkin alters that gather so that runs on greater than 1024 procs will
gather the timing data from 16th of the procs at a time, meaning that it
takes 16 communication steps to get all the data.
This makes that process more communication intensive, but now the 'root' proc
doesn't have to allocate enough space for all procs' timing data. The extra
expense is only a 1-time end-of-run expense.

(From Sierra)