Framework: Change testing default to Experimental
bug: 5551

Changed the default CTEST_TEST_TYPE for all dashboard builds to Experimental.
This was done to reduce the number of tests that are initially sent to the
nightly track, which emails developers, when new testing is set up. The process
that is supposed to be followed is to set up new testing on the experimental
track and then once all tests are passing to promote it to nightly. This change
makes this behavior automatic instead of requiring it to be explicitly stated
in in the test scripts.

This change required most of the existing builds to be updated to explicitly
state that they are nightly. All existing builds that submit to nightly as of
the night of 3/22/12 should continue to submit to the nightly track. The method
I used to update the existing builds was to inspect all the ctest*.cmake scripts
that are called in the CMakeLists.txt for every machine that reported to any of
our dashboards over the previous week. If the ctest script did not have a
CTEST_TEST_TYPE set in it then it was set to "Nightly" as this was where it
would have been reporting to. If one did exist and was set to Experimental then
it was not touched. There were a few exceptions to this. All of the testing on
rouson-mp and s903186 are currently submitting to the experimental branch even
though they did not explicitly set CTEST_TEST_TYPE. This is due to environmental
variables that are set during the run. I added a CTEST_TEST_TYPE and set it to
"Experimental" for all of those builds so that when it is time to promote them
to nightly it will be easier.

There were also a number of ctest scripts that were not in the CMakeLists.txt
for their machine. These scripts were left alone. The reasoning is that since
they aren't run on any schedule it is likely there is some breakage there and
anyone wanting to start running them should start by submitting to experimental
anyway.

There were also several machines that have not been submitting to our dashboard
recently, for example exetazo. I left all of those machines alone for the same
reason as above. We likely need to clean out old build machines that we no
longer use though just to make it more clear what needs to be kept up to date.

I also did not update any of the ORNL machines as I didn't want to cause issues
on their end when I don't have access to their dashboard to see which builds
should be nightly or even if I broke anything.