Explicitly turning on CTEST_TEST_TYPE Nightly now that default is Experimental
The Trilinos commit 94ca1381fb2987587031b02928065aadd9cbbbc3 just made changes
the default of CTEST_TEST_TYPE from Nightly to Experimental.  This was the
right thing to do (and should have been that way all along) but it means that
you now have to explicitly set 'Nightly' in all of the builds that wre
implicitly nightly builds.  I think that I have done this correctly but we
will have to see what happens with testing tomorrow morning.