Repartitioning uses import instead of matmatmult.
Rather than doing Q*R*Ah*P*Q^T, the repartitioning now just uses an importer to
move rows around.  In fact, repartitioning isn't mathematically a permutation.
Instead, it's just a reassignment of rows to different processors, but GIDs stay the same.

Updated the diffusive partition scheme for better control of when it occurs,
see the embedded comments.  Added a cute ASCII visualization of where rows are placed
on a processor grid.

The "Permutation" object has been replaced by an "Importer" on the level.