Xpetra: fixed bug #5603 Bug fixed by keeping internally a reference to the vector returned by getTpetra_Vector()
Description of the problem:
-------------------------------------
- toXpetra() is supposed to return a reference to an internal object stored by
the Xpetra class. The reference is supposed to be valid while the Xpetra object
exist.
- The implementation of Xpetra::Vector is special. Internally, Xpetra::Vector
does not stores a Tpetra::Vector. Instead, it stores a MultiVector. I think
this is to make Xpetra::TpetraVector inherit from Xpetra::TpetraMultiVector.
- toXpetra() returns a reference to an Tpetra::Vector that is created "on
demand" using Tpetra::MultiVector::getVectorNonConst(). This vector is not kept
internally. Therefore, the reference returned by Xpetra is invalid.
- the Epetra version seems OK because Epetra_MultiVector::operator () returns a
vector that is kept internally by the Epetra_MultiVector and freed with the
Epetra_MultiVector.